pyproject.toml
[project]
name = "mlx-embeddings-server"
version = "0.1.0"
description = "FastAPI server for MLX Embeddings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "requests",
    "pillow>=10.2.0",
    "torch",
    "torchvision",
    "mlx",
    "mlx-embeddings @ git+https://github.com/ethanlee928/mlx-embeddings.git@main",
    "mlx-vlm @ git+https://github.com/ethanlee928/mlx-vlm.git@main",
    "colpali-engine==0.3.13"
]

[project.scripts]
mlx-embeddings-server = "mlx_embeddings_server.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
override-dependencies = [
  "transformers==4.57.6",   # for colpali-engine
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "httpx>=0.28.1",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
